refactor: move queries into separate files
