+++ lua/treesitter-context.lua
@@ -2,11 +2,19 @@ local api = vim.api
 local ts_utils = require'nvim-treesitter.ts_utils'
 local highlighter = vim.treesitter.highlighter
 local parsers = require'nvim-treesitter.parsers'
+local util = require'treesitter-context.util'
+local CATEGORY = util.CATEGORY
+local QUERY = util.QUERY
 
 local augroup = api.nvim_create_augroup
 local command = api.nvim_create_user_command
 
-local M = {}
+local M = {
+    CATEGORY = CATEGORY,
+    QUERY = QUERY,
+    field_name_query = util.field_name_query,
+    node_type_query = util.node_type_query,
+}
 
 local defaultConfig = {
   enable = true,
@@ -21,21 +29,6 @@ local defaultConfig = {
 local config = {}
 
 -- Constants
-local CATEGORY = {
-  CLASS = 1,
-  INTERFACE = 2,
-  STRUCT = 3,
-  ENUM = 4,
-  FUNCTION = 5,
-  METHOD = 6,
-  FOR = 7,
-  WHILE = 8,
-  IF = 9,
-  SWITCH = 10,
-  CASE = 11,
-  SECTION = 12,
-}
-
 local DEFAULT_CATEGORIES = {
   default = {
     CATEGORY.CLASS,
@@ -53,443 +46,7 @@ local DEFAULT_CATEGORIES = {
   },
 }
 
-local QUERY = {
-  FIELD_NAME = 1,
-  NODE_TYPE = 2,
-}
-
-function M.field_name_query(name)
-  return {
-    name = name,
-    kind = QUERY.FIELD_NAME,
-  }
-end
-
-function M.node_type_query(name)
-  return {
-    name = name,
-    kind = QUERY.NODE_TYPE,
-  }
-end
-
-local DEFAULT_QUERIES
-do
-  local f = M.field_name_query
-  local t = M.node_type_query
-
-  DEFAULT_QUERIES = {
-    c = {
-      ['struct_specifier'] = {
-        category = CATEGORY.STRUCT,
-        last = { f'name' },
-      },
-      ['enum_specifier'] = {
-        category = CATEGORY.ENUM,
-        last = { f'name' },
-      },
-      ['function_definition'] = {
-        category = CATEGORY.FUNCTION,
-        last = { f'declarator' },
-      },
-      ['for_statement'] = {
-        category = CATEGORY.FOR,
-        last = { f'update', f'condition', f'initializer' },
-      },
-      ['while_statement' ] = {
-        category = CATEGORY.WHILE,
-        last = { f'condition' },
-      },
-      ['if_statement' ] = {
-        category = CATEGORY.IF,
-        last = { f'condition' },
-      },
-      ['switch_statement' ] = {
-        category = CATEGORY.SWITCH,
-        last = { f'condition' },
-      },
-      ['case_statement' ] = {
-        category = CATEGORY.CASE,
-        last = { f'value' },
-      },
-    },
-    cpp = {
-      ['class_specifier'] = {
-        category = CATEGORY.CLASS,
-        last = { t'base_class_clause', f'name' },
-      },
-      ['struct_specifier'] = {
-        category = CATEGORY.STRUCT,
-        last = { t'base_class_clause', f'name' },
-      },
-      ['enum_specifier'] = {
-        category = CATEGORY.ENUM,
-        last = { f'name' },
-      },
-      ['function_definition'] = {
-        category = CATEGORY.FUNCTION,
-        last = { f'declarator' },
-      },
-      ['for_statement'] = {
-        category = CATEGORY.FOR,
-        last = { f'update', f'condition', f'initializer' },
-      },
-      ['while_statement' ] = {
-        category = CATEGORY.WHILE,
-        last = { f'condition' },
-      },
-      ['if_statement' ] = {
-        category = CATEGORY.IF,
-        last = { f'condition' },
-      },
-      ['switch_statement' ] = {
-        category = CATEGORY.SWITCH,
-        last = { f'condition' },
-      },
-      ['case_statement' ] = {
-        category = CATEGORY.CASE,
-        last = { f'value' },
-      },
-    },
-    rust = {
-      ['impl_item'] = {
-        category = CATEGORY.CLASS,
-        last = { f'type' },
-      },
-      ['trait_item'] = {
-        category = CATEGORY.INTERFACE,
-        last = { f'bounds', f'name' },
-      },
-      ['struct_item'] = {
-        category = CATEGORY.STRUCT,
-        last = { f'name' },
-      },
-      ['enum_item'] = {
-        category = CATEGORY.ENUM,
-        last = { f'name' },
-      },
-      ['function_item'] = {
-        category = CATEGORY.FUNCTION,
-        last = { f'return_type', f'parameters' },
-      },
-      ['for_expression'] = {
-        category = CATEGORY.FOR,
-        last = { f'value' },
-      },
-      ['while_expression'] = {
-        category = CATEGORY.WHILE,
-        last = { f'condition' },
-      },
-      ['while_let_expression'] = {
-        category = CATEGORY.WHILE,
-        last = { f'value' },
-      },
-      ['loop_expression'] = {
-        category = CATEGORY.WHILE,
-      },
-      ['if_expression' ] = {
-        category = CATEGORY.IF,
-        last = { f'condition' },
-      },
-      ['if_let_expression' ] = {
-        category = CATEGORY.IF,
-        last = { f'value' },
-      },
-      ['match_expression' ] = {
-        category = CATEGORY.SWITCH,
-        last = { f'value' },
-      },
-      ['match_arm' ] = {
-        category = CATEGORY.CASE,
-        last = { f'pattern' },
-      },
-    },
-    java = {
-      ['class_declaration'] = {
-        category = CATEGORY.CLASS,
-        last = { f'interfaces', f'superclasses', f'name' },
-      },
-      ['enum_declaration'] = {
-        category = CATEGORY.ENUM,
-        last = { f'name' },
-      },
-      ['interface_declaration'] = {
-        category = CATEGORY.INTERFACE,
-        last = { f'extends_interfaces', f'name' },
-      },
-      ['constructor_declaration'] = {
-        category = CATEGORY.METHOD,
-        last = { f'parameters' },
-      },
-      ['method_declaration'] = {
-        category = CATEGORY.METHOD,
-        last = { f'parameters' },
-      },
-      ['for_statement'] = {
-        category = CATEGORY.FOR,
-        last = { f'update', f'condition', f'init' },
-      },
-      ['enhanced_for_statement'] = {
-        category = CATEGORY.FOR,
-        last = { f'value' },
-      },
-      ['while_statement'] = {
-        category = CATEGORY.WHILE,
-        last = { f'condition' },
-      },
-      ['if_statement' ] = {
-        category = CATEGORY.IF,
-        last = { f'condition' },
-      },
-      ['switch_expression' ] = {
-        category = CATEGORY.SWITCH,
-        last = { f'condition' },
-      },
-      ['switch_block_statement_group' ] = {
-        category = CATEGORY.CASE,
-        last = { t'switch_label' },
-      },
-      ['switch_rule' ] = {
-        category = CATEGORY.CASE,
-        last = { t'switch_label' },
-      },
-    },
-    javascript = {
-      ['class_declaration'] = {
-        category = CATEGORY.CLASS,
-        last = { t'class_heritage', f'name' },
-      },
-      ['function_declaration'] = {
-        category = CATEGORY.FUNCTION,
-        last = { f'parameters' },
-      },
-      ['method_declaration'] = {
-        category = CATEGORY.METHOD,
-        last = { f'parameters' },
-      },
-      ['for_statement'] = {
-        category = CATEGORY.FOR,
-        last = { f'increment' },
-      },
-      ['for_in_statement'] = {
-        category = CATEGORY.FOR,
-        last = { f'right' },
-      },
-      ['while_statement'] = {
-        category = CATEGORY.WHILE,
-        last = { f'condition' },
-      },
-      ['if_statement' ] = {
-        category = CATEGORY.IF,
-        last = { f'condition' },
-      },
-      ['switch_statement' ] = {
-        category = CATEGORY.SWITCH,
-        last = { f'value' },
-      },
-      ['switch_case' ] = {
-        category = CATEGORY.CASE,
-        last = { f'pattern' },
-      },
-      ['switch_default' ] = {
-        category = CATEGORY.CASE,
-      },
-    },
-    typesript = {
-      ['class_declaration'] = {
-        category = CATEGORY.CLASS,
-        last = { t'class_heritage', f'name' },
-      },
-      ['interface_declaration'] = {
-        category = CATEGORY.INTERFACE,
-        last = { t'extends_type_clause', f'name' },
-      },
-      ['function_declaration'] = {
-        category = CATEGORY.FUNCTION,
-        last = { f'return_type', f'parameters' },
-      },
-      ['method_declaration'] = {
-        category = CATEGORY.METHOD,
-        last = { f'return_type', f'parameters' },
-      },
-      ['for_statement'] = {
-        category = CATEGORY.FOR,
-        last = { f'increment' },
-      },
-      ['for_in_statement'] = {
-        category = CATEGORY.FOR,
-        last = { f'right' },
-      },
-      ['while_statement'] = {
-        category = CATEGORY.WHILE,
-        last = { f'condition' },
-      },
-      ['if_statement' ] = {
-        category = CATEGORY.IF,
-        last = { f'condition' },
-      },
-      ['switch_statement' ] = {
-        category = CATEGORY.SWITCH,
-        last = { f'value' },
-      },
-      ['switch_case' ] = {
-        category = CATEGORY.CASE,
-        last = { f'pattern' },
-      },
-      ['switch_default' ] = {
-        category = CATEGORY.CASE,
-      },
-    },
-    lua = {
-      ['function_declaration' ] = {
-        category = CATEGORY.FUNCTION,
-        last = { f'parameters' },
-      },
-      ['for_statement'] = {
-        category = CATEGORY.FOR,
-        last = { f'clause' },
-      },
-      ['while_statement'] = {
-        category = CATEGORY.WHILE,
-        last = { f'condition' },
-      },
-      ['if_statement'] = {
-        category = CATEGORY.IF,
-        last = { f'condition' },
-      },
-    },
-    teal = {
-      ['record_declaration' ] = {
-        category = CATEGORY.CLASS,
-        last = { f'name' },
-      },
-      ['enum_declaration' ] = {
-        category = CATEGORY.ENUM,
-        last = { f'name' },
-      },
-      ['function_statement' ] = {
-        category = CATEGORY.FUNCTION,
-        last = { f'signature' },
-      },
-      ['numeric_for_statement'] = {
-        category = CATEGORY.FOR,
-        last = { f'step', f'target' },
-      },
-      ['generic_for_statement'] = {
-        category = CATEGORY.FOR,
-        last = { f'iterator' },
-      },
-      ['while_statement'] = {
-        category = CATEGORY.WHILE,
-        last = { f'condition' },
-      },
-      ['if_statement'] = {
-        category = CATEGORY.IF,
-        last = { f'condition' },
-      },
-    },
-    python = {
-      ['class_definition'] = {
-        category = CATEGORY.CLASS,
-        last = { f'superclasses', f'name' },
-      },
-      ['function_definition' ] = {
-        category = CATEGORY.FUNCTION,
-        last = { f'return_type', f'parameters' },
-      },
-      ['for_statement'] = {
-        category = CATEGORY.FOR,
-        last = { f'right' },
-      },
-      ['while_statement'] = {
-        category = CATEGORY.WHILE,
-        last = { f'condition' },
-      },
-      ['if_statement'] = {
-        category = CATEGORY.IF,
-        last = { f'condition' },
-      },
-      ['with_statement'] = {
-        category = CATEGORY.IF,
-        last = { f'with_clause' },
-      },
-      ['match_statement'] = {
-        category = CATEGORY.SWITCH,
-        last = { f'subject' },
-      },
-      ['case_clause'] = {
-        category = CATEGORY.CASE,
-        last = { f'pattern' },
-      },
-    },
-    php = {
-      ['class_declaration'] = {
-        category = CATEGORY.CLASS,
-        skip = { f'attributes' },
-        last = { t'class_interface_clause', t'base_clause', f'name' },
-      },
-      ['interface_declaration'] = {
-        category = CATEGORY.INTERFACE,
-        last = { t'base_clause', f'name' },
-      },
-      ['enum_declaration'] = {
-        category = CATEGORY.ENUM,
-        last = { f'name' },
-      },
-      ['method_declaration'] = {
-        category = CATEGORY.METHOD,
-        skip = { f'attributes' },
-        last = { f'return_type', f'parameters' },
-      },
-      ['for_statement'] = {
-        category = CATEGORY.FOR,
-      },
-      ['foreach_statement'] = {
-        category = CATEGORY.FOR,
-      },
-      ['while_statement'] = {
-        category = CATEGORY.WHILE,
-        last = { f'condition' },
-      },
-      ['if_statement'] = {
-        category = CATEGORY.IF,
-        last = { f'condition' },
-      },
-      ['switch_statement'] = {
-        category = CATEGORY.SWITCH,
-        last = { f'condition' },
-      },
-      ['case_statement'] = {
-        category = CATEGORY.CASE,
-        last = { f'value' },
-      },
-      ['default_statement'] = {
-        category = CATEGORY.CASE,
-      },
-    },
-    tex = {
-      ['part'] = {
-        category = CATEGORY.SECTION,
-        last = { f'text' },
-      },
-      ['chapter'] = {
-        category = CATEGORY.SECTION,
-        last = { f'text' },
-      },
-      ['section'] = {
-        category = CATEGORY.SECTION,
-        last = { f'text' },
-      },
-      ['subsection'] = {
-        category = CATEGORY.SECTION,
-        last = { f'text' },
-      },
-      ['subsubsection'] = {
-        category = CATEGORY.SECTION,
-        last = { f'text' },
-      },
-    },
-  }
-end
+local DEFAULT_QUERIES = require'treesitter-context.queries'
 
 local DEFAULT_FALLBACK_PATTERNS = {
    -- These catch most generic groups, eg "function_declaration" or "function_block"
@@ -773,8 +330,6 @@ local function display_window(bufnr, winid, width, height, col, ty, hl)
 end
 
 -- Exports
-
-M.config = config
 
 local function get_parent_matches(max_lines)
   if max_lines == 0 then
