+++ lua/treesitter-context/queries.lua
@@ -1,0 +1,13 @@
+return {
+  c = require('treesitter-context.queries.c'),
+  cpp = require('treesitter-context.queries.cpp'),
+  rust = require('treesitter-context.queries.rust'),
+  java = require('treesitter-context.queries.java'),
+  javascript = require('treesitter-context.queries.javascript'),
+  typescript = require('treesitter-context.queries.typescript'),
+  lua = require('treesitter-context.queries.lua'),
+  teal = require('treesitter-context.queries.teal'),
+  python = require('treesitter-context.queries.python'),
+  php = require('treesitter-context.queries.php'),
+  tex = require('treesitter-context.queries.tex'),
+}
